🎨 [Refactor] 이벤트 생성 기능 관리자로 분리 #64
