+++ src/main/java/com/devcv/admin/application/AdminService.java
@@ -1,5 +1,8 @@
 package com.devcv.admin.application;
 
+import com.devcv.event.domain.Event;
+import com.devcv.event.domain.dto.EventRequest;
+import com.devcv.event.repository.EventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -7,4 +10,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class AdminService {
 
+    private final EventRepository eventRepository;
+
+    public Event createEvent(EventRequest eventRequest) {
+        Event event = Event.of(eventRequest.name(), eventRequest.startDate(), eventRequest.endDate());
+        return eventRepository.save(event);
+    }
 }
