+++ src/main/java/com/devcv/admin/presentation/AdminController.java
@@ -1,9 +1,11 @@
 package com.devcv.admin.presentation;
 
+import com.devcv.admin.application.AdminService;
 import com.devcv.auth.application.AuthService;
-import com.devcv.auth.details.MemberDetails;
 import com.devcv.auth.exception.JwtInvalidSignException;
 import com.devcv.common.exception.ErrorCode;
+import com.devcv.event.domain.Event;
+import com.devcv.event.domain.dto.EventRequest;
 import com.devcv.member.domain.dto.MemberLoginRequest;
 import com.devcv.member.domain.dto.MemberLoginResponse;
 import com.devcv.resume.application.ResumeService;
@@ -11,21 +13,23 @@ import com.devcv.resume.domain.enumtype.ResumeStatus;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
 
 @RestController
 @RequestMapping("/admin/*")
 @AllArgsConstructor
 public class AdminController {
 
-    private final AuthService authService;
-    private final ResumeService resumeService;
     public static final String AUTHORIZATION_HEADER = "Authorization";
     public static final String AUTHORIZATION_REFRESH_HEADER = "RefreshToken";
     public static final String BEARER_PREFIX = "Bearer ";
+
+    private final AuthService authService;
+    private final ResumeService resumeService;
+    private final AdminService adminService;
+
 
     //----------- login start -----------
     @PostMapping("/login")
@@ -46,5 +50,11 @@ public class AdminController {
     public ResponseEntity<?> adminUpdateResumeStatus(@PathVariable Long resumeId,@PathVariable ResumeStatus status) {
         resumeService.updateStatus(resumeId, status);
         return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/events")
+    public ResponseEntity<Object> createEvent(@RequestBody EventRequest eventRequest) {
+        Event event = adminService.createEvent(eventRequest);
+        return ResponseEntity.created(URI.create(String.valueOf(event.getId()))).build();
     }
 }
