+++ src/main/java/com/devcv/event/application/EventService.java
@@ -4,7 +4,6 @@ import com.devcv.common.exception.ErrorCode;
 import com.devcv.common.exception.NotFoundException;
 import com.devcv.event.domain.Event;
 import com.devcv.event.domain.dto.EventListResponse;
-import com.devcv.event.domain.dto.EventRequest;
 import com.devcv.event.domain.dto.EventResponse;
 import com.devcv.event.repository.EventRepository;
 import lombok.RequiredArgsConstructor;
@@ -17,11 +16,6 @@ import java.util.List;
 public class EventService {
 
     private final EventRepository eventRepository;
-
-    public Event createEvent(EventRequest eventRequest) {
-        Event event = Event.of(eventRequest.name(), eventRequest.startDate(), eventRequest.endDate());
-        return eventRepository.save(event);
-    }
 
     public Event findByEventId(Long eventId) {
         return eventRepository.findById(eventId)
